requirements.txt
python-pptx
tqdm
PyMuPDF